+++ depthview/depth_viewer.py
@@ -3,7 +3,7 @@ A tool to view depth(as npy file) and left image.
 In --disp3d case, use open3d to show 3D point cloud.
 """
 
-from depthview.util_depth_view import view_by_colormap, view3d
+from depthview.depth_view import view_by_colormap, view3d
 
 
 def main():
